### 2020-05-21 #2188 @arfon
Whedon is struggling with compiling the paper here as the repository is very large (~500Mb) - it might be a good idea to look at reducing this @agahkarakuzu. For now, I have compiled the paper locally: [10.21105.joss.02188.pdf](https://github.com/openjournals/joss-reviews/files/4662829/10.21105.joss.02188.pdf)

👋 @cMadan - would you be able to take this paper? Ariel would be the ideal editor but he's unable to edit for JOSS right now.
