### 2020-05-21 #2188 @agahkarakuzu
Hi @arfon, thank you for the quick response and the feedback!

The repo had some large files in the past, hence the large size with deep clone depth. Right now the documentation files are stored in this repo, moving them elsewhere would free some considerable space, but the same issue would still persist with clone.  Do you have a suggestion how we might approach this?

We discussed with the team about removing the large files from the history, but it looked like a tricky operation, so we did not act on it.
