### 2020-05-21 #2188 @mathieuboudreau
@arfon (Hi, one of @agahkarakuzu 's coauthors here) Does Whedon clone the entire repo's git history, or just the master branch and/or a single commit of it?
