### 2020-05-21 #2188 @arfon
> @arfon (Hi, one of @agahkarakuzu 's coauthors here) Does Whedon clone the entire repo's git history, or just the master branch and/or a single commit of it?

It does a full clone. I could investigate how to do a shallow clone.
