### 2020-05-22 #2188 @arfon
I've done a little digging here and it turns out we need the full history for calculating things like the repository statistics.

> We discussed with the team about removing the large files from the history, but it looked like a tricky operation, so we did not act on it.

This is the only thing that will truly fix this issue but it's not essential for the JOSS review as I can compile the paper locally and update the thread with the PDF as we go.

That said, I would encourage you to address this issue as all of your future users will have to clone a very large repository which isn't a great user experience.
