### 2020-05-29 #2188 @Kevin-Mattheus-Moerman
@mfroeling, @oliverchampion, @grlee77, @JoshKarpel, @rljacobson, @angielaird, @Lestropie would you be interested in reviewing: _qMRLab: Quantitative MRI Analysis, under one umbrella_, for [JOSS (the Journal of Open Source Software)](joss.theoj.org/).
The review takes place on GitHub and focuses on [the software](https://github.com/qMRLab/qMRLab) and this short paper:
https://github.com/openjournals/joss-reviews/files/4662829/10.21105.joss.02188.pdf

If you agree to review this work you can let me know here and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issues.

Thanks!
