### 2020-06-03 #2188 @mriedel56
Hi @Kevin-Mattheus-Moerman, I dont think I’m knowledgeable enough about qMRI to provide a sufficient review; but happy to be considered for other submissions.
