### 2020-06-03 #2188 @Lestropie
While I would like to do so and familiarise myself with JOSS prior to hopefully submitting my own work there in the future, I unfortunately cannot commit to a thorough review right now due to other obligations. But thank you for the offer.
