### 2020-06-19 75acc9e
Add b1_afi model
