### 2020-06-23 126a14e
Add filter class inheritance, refactor code for compatibility. Remove parallelize to make it voxelwise = 0.
