### 2020-06-23 0b54cf2
Change default filter to median
