+++ src/Common/FilterClass.m
@@ -19,7 +19,7 @@ classdef (Abstract) FilterClass
     properties
         % Model options
         buttons ={'PANEL','Smoothing filter',6,...
-            'Type',{'polynomial','gaussian','median','spline'},...
+            'Type',{'median','gaussian','spline', 'polynomial'},...
             'Dimension',{'3D','2D'},...
             'size x',3,...
             'size y',3,...
