### 2020-06-23 1aa90ef
Add assumptions, rename data files of CLI usage
