### 2020-06-24 cad4e47
Update b1_afi.m
