### 2020-06-24 bebf159
Add static model
