### 2020-07-02 f7ecb99
Implemented rudimentary general equation fitting
