### 2020-07-02 78295f3
Refactor options to GUI, resolve failing test
