### 2020-07-02 465abf9
Implemented complex fitting
