### 2020-07-02 #2343 @grlee77
Re: **installation instructions** item:

Your [documentation for this](https://qmrlab.readthedocs.io/en/master/install.html) is good, but it would be nice if you also listed which commercial Matlab toolboxes are required. I saw that for Octave, specific toolboxes were listed.

FWIW, I ran a few of the demos and then typed `license('inuse')` and saw
```
image_toolbox
matlab
optimization_toolbox
```
but I did not run everything, so there may be others as well.
