### 2020-07-02 #2343 @grlee77
Please double-check the formatting of the references. For example the first reference (SPM) seems to be incomplete and the Weiskopf et. al. reference has a stray "Citeseer" listed.
