### 2020-07-02 #2343 @grlee77
Re: **citations**:

The authors cite several relevant packages, but one area that does not seem to be represented is for Diffusion MRI. qMRLab does provide DTI, CHARMED and NODDI functionality, so it seems relevant to cite some packages in this area as well. Some common tools for that are free, but not open source, but a few open source ones I am aware of are:

[DIPY](https://dipy.org) and [DMIPY](https://github.com/AthenaEPI/dmipy) (Python)  (full disclosure: I have been an occasional DIPY contributor)
[Camino](http://camino.cs.ucl.ac.uk/) (Java)
[DSI Studio](http://dsi-studio.labsolver.org/) (C++)

Of course, it will not be possible to cite everything, but 1-2 sentences regarding other open-source diffusion processing tools would be nice.
