### 2020-07-03 #2343 @Kevin-Mattheus-Moerman
> Regarding the included figure, it is pretty, but I wonder if there is any issue on JOSS's end to publish a figure including logos of other packages like Matlab, Docker, Plotly, etc. When we submitted the SciPy paper (albeit not for JOSS), there were originally various logos in a historical timeline figure, but the editorial team required the removal of any copyrighted logos prior to publication.
>
> Also, when performing this review, I have only reviewed qMRLab itself, but did not evaluate the related projects like qMRFlow, etc. I think it is appropriate to mention these related projects in the text, but I don't think the figure needs to include them. I guess what I am saying is that the figure looks a bit more like an advertisement for the broader software eco-system surrounding qMRLab and should perhaps be a bit more narrowly focused on the quantitative mapping subset described in the publication.
>
> @Kevin-Mattheus-Moerman: can you advise on any potential issues regarding logos on JOSS's end?

@grlee77 thanks for raising this issue with us. The logos are acceptable.
