### 2020-07-06 #2343 @62442katieb
> Re: **Contributions and authorship** item
>
> Can the authors clarify briefly how authorship criteria were decided? I see that most authors overlap with the repository's significant source code contributors, but I understand contributions can also come in other forms. Two authors I don't see an obvious alias for in the commit history are Drs. Pike and Stikov. Also, there appear to be a few significant contributors (e.g. 30+ commits) who do not appear in the author list (although it is hard to tell for sure based on GitHub handles).

I would also like clarification on the authorship decisions. Thanks!
