### 2020-07-06 #2343 @agahkarakuzu
@grlee77 @62442katieb

Re: Contributions and authorship item

> Two authors I don't see an obvious alias for in the commit history are Drs. Pike and Stikov.

* Bruce Pike and Nikola Stikov (my advisor) are principal investigators of the research labs funding our project

> Can the authors clarify briefly how authorship criteria were decided?

* For the authorship, we thought maintainers, active contributors and PIs fit for the authorship. These are the contributors regularly joining our bi-weekly meetings. For clarity,  below is a brief breakdown of authorship roles:

     * Agah Karakuzu: Maintainer, improved the qMRLab framework, significant contributions to the codebase, added DevOps and new qMRI models, wrote the manuscript.
     * Mathieu Boudreau: Maintainer, significant contributions to the codebase, added new qMRI models, revised the manuscript
     * Tanguy Duval: Former maintainer, created qMRLab framework based on qMTLab, significant contributions to the codebase, revised the manuscript
     * Tommy Boshkovski: Active contributor, added qMRI models, helped improve documentation, created the web portal, revised the manuscript.
     * Ilana R. Leppert: Active contributor, added qMRI models, revised the manuscript.
     * Bruce Pike: Principal Investigator, contributed funds for the development of the software, revised the manuscript.
     * Julien Cohen-Adad: Principal Investigator, contributions to the codebase,  contributed funds for the development of the software, provided valuable insights on software development, revised the manuscript.
     * Nikola Stikov: Principal Investigator, main supervisor of the project and academic supervisor of A.K. and T.B., contributed funds for the development of the software, revised the manuscript

> Also, there appear to be a few significant contributors (e.g. 30+ commits) who do not appear in the author list (although it is hard to tell for sure based on GitHub handles).

* I would like to note that later on in the project, we started using `squash and merge` option to merge the commits into the `master`. Therefore, at least for this codebase, number of commits is not necessarily an objective measure of the relative code contributions. Nonetheless, we are more than happy to welcome all the contributors as authors if the JOSS policy (@Kevin-Mattheus-Moerman ) requires and/or the reviewers @grlee77 @62442katieb strongly recommend to do so.

> Also, when performing this review, I have only reviewed qMRLab itself, but did not evaluate the related projects like qMRFlow, etc. I think it is appropriate to mention these related projects in the text, but I don't think the figure needs to include them.

Thank you @grlee77, I agree with your point. I will update the figure to confine the content to qMRLab codebase only and simplify it.

- [ ] The comment will be addressed in the manuscript.

> The authors cite several relevant packages, but one area that does not seem to be represented is for Diffusion MRI. qMRLab does provide DTI, CHARMED and NODDI functionality, so it seems relevant to cite some packages in this area as well.

Thank you for the suggestion. I agree that including some open source examples from the Diffusion MRI field is relevant.

- [ ] The comment will be addressed in the manuscript.

- [ ]  Please double-check the formatting of the references.

- [ ]  it would be nice if you also listed which commercial Matlab toolboxes are required

Thank you so much @grlee77 for your valuable comments, I will soon address them and make a PR.
