### 2020-07-06 ff59c18
[ADD] MATLAB Toolbox Dependencies
* Add information about required MATLAB tools 
* Addressing the suggestion by @grlee77 https://github.com/openjournals/joss-reviews/issues/2343#issuecomment-653094720
