### 2020-07-06 88f3d27
Address several comments by @grlee77
* Added references to dMRI packages to better represent the field https://github.com/openjournals/joss-reviews/issues/2343#issuecomment-653122946
* Fixed reference formatting https://github.com/openjournals/joss-reviews/issues/2343#issuecomment-653119291
* Narrowed down the context of the figure only to qMRLab component https://github.com/openjournals/joss-reviews/issues/2343#issuecomment-653129589
* Logos of the other software were not removed as @Kevin-Mattheus-Moerman noticed that the logos are acceptable
