+++ paper.bib
@@ -29,10 +29,15 @@
   publisher={Elsevier}
 }
 
-@article{ashburner:1994,
-  title={Statistical parametric mapping},
-  author={Ashburner, John and Barnes, Gareth and Chen, Chun-Chuan and Daunizeau, Jean and Flandin, Guillaume and Friston, Karl and Glauche, Volkmar and Henson, Rik and Hutton, Chloe and Kiebel, Stephan and others},
-  year={1994}
+@article{ashburner:2012,
+  title={SPM: a history},
+  author={Ashburner, John},
+  journal={Neuroimage},
+  volume={62},
+  number={2},
+  pages={791--800},
+  year={2012},
+  publisher={Elsevier}
 }
 
 @inproceedings{weiskopf:2008,
@@ -41,8 +46,7 @@
   booktitle={Proceedings of ISMRM},
   volume={16},
   pages={2241},
-  year={2008},
-  organization={Citeseer}
+  year={2008}
 }
 
 @article{hafyane:2018,
@@ -91,4 +95,54 @@
   pages={23--49},
   year={1999},
   publisher={Springer}
+}
+
+@article{di:2017,
+  title={Nextflow enables reproducible computational workflows},
+  author={Di Tommaso, Paolo and Chatzou, Maria and Floden, Evan W and Barja, Pablo Prieto and Palumbo, Emilio and Notredame, Cedric},
+  journal={Nature biotechnology},
+  volume={35},
+  number={4},
+  pages={316--319},
+  year={2017},
+  publisher={Nature Publishing Group}
+}
+
+@article{theaud:2020,
+  title={TractoFlow: A robust, efficient and reproducible diffusion MRI pipeline leveraging Nextflow \& Singularity},
+  author={Theaud, Guillaume and Houde, Jean-Christophe and Bor{\'e}, Arnaud and Rheault, Fran{\c{c}}ois and Morency, Felix and Descoteaux, Maxime},
+  journal={NeuroImage},
+  pages={116889},
+  year={2020},
+  publisher={Elsevier}
+}
+
+@inproceedings{cook:2006,
+  title={Camino: open-source diffusion-MRI reconstruction and processing},
+  author={Cook, PA and Bai, Yu and Nedjati-Gilani, SKKS and Seunarine, KK and Hall, MG and Parker, GJ and Alexander, D Cq},
+  booktitle={14th scientific meeting of the international society for magnetic resonance in medicine},
+  volume={2759},
+  pages={2759},
+  year={2006},
+  organization={Seattle WA, USA}
+}
+
+@article{garyfallidis:2014,
+  title={Dipy, a library for the analysis of diffusion MRI data},
+  author={Garyfallidis, Eleftherios and Brett, Matthew and Amirbekian, Bagrat and Rokem, Ariel and Van Der Walt, Stefan and Descoteaux, Maxime and Nimmo-Smith, Ian},
+  journal={Frontiers in neuroinformatics},
+  volume={8},
+  pages={8},
+  year={2014},
+  publisher={Frontiers}
+}
+
+@article{fick:2019,
+  title={The dmipy toolbox: Diffusion mri multi-compartment modeling and microstructure recovery made easy},
+  author={Fick, Rutger HJ and Wassermann, Demian and Deriche, Rachid},
+  journal={Frontiers in neuroinformatics},
+  volume={13},
+  pages={64},
+  year={2019},
+  publisher={Frontiers}
 }
