### 2020-07-06 #2343 @agahkarakuzu
@Kevin-Mattheus-Moerman I responded to the comments by @grlee77 and @62442katieb and addressed the issues raised in [qMRLab PR #399](https://github.com/qMRLab/qMRLab/pull/399) ([`joss_rev1`](https://github.com/qMRLab/qMRLab/pull/399) into `master`).
