### 2020-07-06 #2343 @grlee77
> I would like to note that later on in the project, we started using squash and merge option to merge the commits into the master. Therefore, at least for this codebase, number of commits is not necessarily an objective measure of the relative code contributions. Nonetheless, we are more than happy to welcome all the contributors as authors if the JOSS policy (@Kevin-Mattheus-Moerman ) requires and/or the reviewers @grlee77 @62442katieb strongly recommend to do so.

Agreed, Re: commits not being an ideal measure of effort. I'm not sure in regards to JOSS policy, but given that it only seems to be a few additional contributors, it shouldn't be much work to invite them. I prefer to lean toward the side of being more inclusive rather than less. If they are uninterested or you do not receive a timely response, then do not include them among the primary authors, perhaps just adding a generic acknowledgement of the contributions of "qMRLab community contributors".
