### 2020-07-07 #2343 @Kevin-Mattheus-Moerman
@agahkarakuzu @grlee77
In relation to authorship here is our policy on this: https://joss.readthedocs.io/en/latest/submitting.html#authorship. In short _"The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed."_,
In this case some contributors were identified which perhaps could be added as authors. I leave this decision with @agahkarakuzu. If this concerns a small amount of contributors, and you do not mind, then I do suggest inviting them (if their contributions are sufficient enough) to see if they feel they would like to be added as a co-author here.
