### 2020-07-07 #2343 @grlee77
It looks like qMRLab/qMRLab#399 addresses all of my other comments (the new figure looks good!). Thanks @agahkarakuzu for the quick update.
