### 2020-07-08 #2343 @agahkarakuzu
@grlee77 regarding authorship, we will send all the contributors an email and ask if they would like to be co-author. We will wait for a week to collect the responses and the necessary information, then update the author list.
