### 2020-07-08 23139af
Add functions for Bloch simulations and analytical equation
