### 2020-07-08 a409599
Add function for the analytical equation
