### 2020-07-08 564d342
Calculate ra/rb for general model; return idx and res as well
