+++ src/Models/T1_relaxometry/inversion_recovery.m
@@ -206,12 +206,17 @@ end
                         end
                         [~,ind] = min(resnorm); % Index of the minimum residual will be which signal fit results to choose.
                         FitResults.T1 = fitVals{ind}.T1;
-                        FitResults.resnorm = resnorm(ind);
+                        FitResults.ra = fitVals{ind}.ra;
+                        FitResults.rb = fitVals{ind}.rb;
+                        FitResults.res = resnorm(ind);
+                        if (strcmp(obj.options.method, 'Magnitude'))
+                            FitResults.idx = ind;
+                        end
                     elseif strcmp(obj.options.method, 'Complex')
                         params.dataType = 'complex';
                         [fitVals, resnorm] = inversion_recovery.fit_lm(data(:), params, 3);
                         FitResults.T1 = fitVals.T1;
-                        FitResults.resnorm = resnorm;
+                        FitResults.res = resnorm;
                     end
             end
         end
@@ -527,6 +532,16 @@ end
                         [x, resnorm] = lsqnonlin(@(x)ir_loss_func_3(x, TR, TI, dataNorm(:)'), x0, [0, 0], [2, 5000], options);
 
                         fitVals.T1 = x(2);
+                        
+                        % ra and rb calculation from Equation 3
+                        if max(abs(data)) > 0
+                            fitVals.ra = x(1)*max(abs(data)) * (1 + exp(-TR/fitVals.T1));
+                            fitVals.rb = -2*x(1)*max(abs(data));
+                        else
+                            fitVals.ra = x(2) * (1 + exp(-TR/fitVals.T1));
+                            fitVals.rb = -2*x(2);
+                        end
+
                     end
 
 
