### 2020-07-08 b69168f
Fixing a parenthesis grouping error on equations
