### 2020-07-09 2cc53ae
Save ra and rb for complex dataType during General equation fit
