### 2020-07-12 #2343 @Kevin-Mattheus-Moerman
@grlee77 just had a look at your unchecked boxes. "References" and "Contribution/authorship" are being dealt with. Can you summarize why "State of the field" is unchecked? Thanks!
