### 2020-07-12 #2343 @grlee77
> @grlee77 just had a look at your unchecked boxes. "References" and "Contribution/authorship" are being dealt with. Can you summarize why "State of the field" is unchecked? Thanks!

State of the field was also addressed in their PR (discussion of existing diffusion-related software was added). I wasn't sure if I should wait until it was merged to check them off, but have gone ahead and checked them now so I don't hold up the review.
