### 2020-07-15 #2343 @agahkarakuzu
@grlee77 I sent contributors an email and waited for a week for their response. Now I will add those accepted co-authorship in the PR and then merge.
