### 2020-07-15 7f61d5b
ADD missing DOIs manually
