+++ paper.bib
@@ -16,6 +16,7 @@
   volume={3},
   number={26},
   pages={656},
+  doi={10.21105/joss.00656},
   year={2018}
 }
 
@@ -37,6 +38,7 @@
   number={2},
   pages={791--800},
   year={2012},
+  doi = {10.1016/b978-012372560-8/50002-4},
   publisher={Elsevier}
 }
 
@@ -55,6 +57,7 @@
   journal={BioRxiv},
   pages={343079},
   year={2018},
+  doi = {10.1101/343079},
   publisher={Cold Spring Harbor Laboratory}
 }
 
@@ -65,6 +68,7 @@
   volume={11},
   pages={100369},
   year={2020},
+  doi={10.1016/j.softx.2019.100369},
   publisher={Elsevier}
 }
 
@@ -94,6 +98,7 @@
   number={3},
   pages={23--49},
   year={1999},
+  doi={10.4135/9781412950657.n33},
   publisher={Springer}
 }
 
