### 2020-07-15 #2343 @agahkarakuzu
@Kevin-Mattheus-Moerman I added missing dois manually (added `doi` field(s) in `paper.bib`), I checked them one by one and they all resolve correctly. Now they appear in the refs:

![image](https://user-images.githubusercontent.com/9632322/87534488-16d04a80-c664-11ea-9858-6c277f488d62.png)

If this looks OK, I will merge the PR to master.
