### 2020-07-15 e51d316
Joss rev1 (#399)
* [ADD] MATLAB Toolbox Dependencies 

* Add information about required MATLAB tools 
* Addressing the suggestion by @grlee77 https://github.com/openjournals/joss-reviews/issues/2343#issuecomment-653094720

* Address several comments by @grlee77

* Added references to dMRI packages to better represent the field https://github.com/openjournals/joss-reviews/issues/2343#issuecomment-653122946
* Fixed reference formatting https://github.com/openjournals/joss-reviews/issues/2343#issuecomment-653119291
* Narrowed down the context of the figure only to qMRLab component https://github.
[…]
