### 2020-07-20 #2343 @mfroeling
@agahkarakuzu My appologies for the delay. I was a bit bussy and needed to find a matlab version with the correct toolboxes which in our IT department had a bit of a delay.

some issues that I think need to be solved.

- [ ] When the toolbox starts it needs a check if the needed Matlab dependencies are installed. The first few tries I had issues during data processing that were due to missing or badly installed packages. Especially for the missing packages, there should be a strong warning when starting the toolbox that it will not run properly.

- [ ] Example data is provided for the T1 relaxometry in the documentation. After some browsing the GUI I found the download example button also for other modalities. This should be better documented. I started with the command line and there I could not find a list of available example data.

- [ ] Each model needs a specific data setup that is apparently inferred from the files in the folder. One can load each data needed separately or by selecting the containing folder. However, for me, it is not clear how the data structure/naming should be for every type of data supported if one wants to used the predefined data format. For example, when additional data quantitative data is loaded how should this be stored (e.g. B1 can be fraction or percent).

- [ ] I loaded one of my own diffusion sets and the bval/bvec files were not loaded or asked for. without this information the tool still allowed me to click the fit button resulting in errors. Only after some exploration I found the load bval/bvec in the options pannel. Nowhere in the help of the diffusion does it state that I have to load my own bval/bvec files.

- [ ] different models ask for different input files. This is nicely provided in the GUI however I would be nice if mandatory and optional inputs would be clear. For example, with DTI processing I don't need a mask or SigmaNoise to be able to fit a Tensor.

- [ ] When I loaded the example diffusion data and then changed the input to one of my own datasets the mask of the previous dataset remained in the Mask input field only after clicking browse and cancel is was cleared.
