### 2020-07-22 #2343 @Kevin-Mattheus-Moerman
@agahkarakuzu your paper compiled now but it looks like some DOI's are still missing :point_up: can you check/add those?
