### 2020-07-28 3c1bd2b
Improve doc & GUI help
* Supplementary information re example datsets
* Add help button for prots that appears when tabletip is passed 
* Add tabletip for all the diffusion models.
