### 2020-07-28 6c14248
Improve req/opt conditions and input types
* Update B1map definitions 
* Make info appear in the textbox
