+++ src/Models/T1_relaxometry/vfa_t1.m
@@ -4,9 +4,11 @@ classdef vfa_t1 < AbstractModel
 % Assumptions:
 %
 % Inputs:
-%   VFAData         spoiled Gradient echo data, 4D volume with different flip angles in time dimension
-%   (B1map)         excitation (B1+) fieldmap. Used to correct flip angles. (optional)
-%   (Mask)          Binary mask to accelerate the fitting (optional)
+%   VFAData         Spoiled Gradient echo data, 4D volume with different flip angles in time dimension
+%   (B1map)         Normalized transmit excitation field map (B1+). B1+ is defined 
+%                   as a  normalized multiplicative factor such that:
+%                   FA_actual = B1+ * FA_nominal. (OPTIONAL).
+%   (Mask)          Binary mask to accelerate the fitting. (OPTIONAL)
 %
 % Outputs:
 %   T1              Longitudinal relaxation time [s]
