+++ src/Common/tools/FileBrowser/BrowserSet.m
@@ -74,8 +74,10 @@ classdef BrowserSet
                 Position = [Location + [0.27, 0], 0.58, 0.1];
                 obj.FileBox = uicontrol(obj.parent, 'Style', 'text','units', 'normalized', 'fontunits', 'normalized', 'Position', Position,'FontSize', 0.6,...
                     'BackgroundColor', [1 1 1]);
-                if  InputOptional, set(obj.FileBox,'string','OPTIONAL'); end
-                if ~InputOptional, set(obj.FileBox,'string','REQUIRED'); end
+                
+                if InputOptional && ~isempty(info), set(obj.FileBox,'string',info); end
+                if InputOptional && isempty(info), set(obj.FileBox,'string','OPTIONAL'); end
+                if ~InputOptional, set(obj.FileBox,'string',['REQUIRED ' info]); end
 
                 % add View button
                 Position = [Location + [0.87, 0], 0.10, 0.1];
