+++ src/Models/Magnetization_transfer/qmt_spgr.m
@@ -8,10 +8,12 @@ classdef qmt_spgr < AbstractModel
 %
 % Inputs:
 %   MTdata              Magnetization Transfert data
-%   (R1map)             1/T1map (VFA RECOMMENDED Boudreau 2017 MRM)
-%   (B1map)             B1 field map, used for flip angle correction (=1 if not provided)
-%   (B0map)             B0 field map, used for offset correction (=0Hz if not provided)
-%   (Mask)              Binary mask to accelerate the fitting
+%   (R1map)             1/T1map (VFA RECOMMENDED Boudreau 2017 MRM) (OPTIONAL)
+%   (B1map)             Normalized transmit excitation field map (B1+). B1+ is defined 
+%                       as a  normalized multiplicative factor such that:
+%                       FA_actual = B1+ * FA_nominal. (OPTIONAL). (=1 if not provided)
+%   (B0map)             B0 field map, used for offset correction  (OPTIONAL) (=0Hz if not provided)
+%   (Mask)              Binary mask to accelerate the fitting (OPTIONAL)
 %
 % Outputs:
 %   F                   Ratio of number of restricted pool to free pool, defined
