### 2020-07-28 #2343 @agahkarakuzu
- [x] Add missing DOIs (@Kevin-Mattheus-Moerman)

Comments by @mfroeling
- [x]  When the toolbox starts it needs a check if the needed Matlab dependencies are installed. The first few tries I had issues during data processing that were due to missing or badly installed packages. Especially for the missing packages, there should be a strong warning when starting the toolbox that it will not run properly.

- [x]  Example data is provided for the T1 relaxometry in the documentation. After some browsing the GUI I found the download example button also for other modalities. This should be better documented. I started with the command line and there I could not find a list of available example data.

- [x]  Each model needs a specific data setup that is apparently inferred from the files in the folder. One can load each data needed separately or by selecting the containing folder. However, for me, it is not clear how the data structure/naming should be for every type of data supported if one wants to used the predefined data format. For example, when additional data quantitative data is loaded how should this be stored (e.g. B1 can be fraction or percent).

- [x]  I loaded one of my own diffusion sets and the bval/bvec files were not loaded or asked for. without this information the tool still allowed me to click the fit button resulting in errors. Only after some exploration I found the load bval/bvec in the options pannel. Nowhere in the help of the diffusion does it state that I have to load my own bval/bvec files.

- [ ]  different models ask for different input files. This is nicely provided in the GUI however I would be nice if mandatory and optional inputs would be clear. For example, with DTI processing I don't need a mask or SigmaNoise to be able to fit a Tensor.

I will open a PR on `qMRLab/qMRLab` repo (`joss_rev2`-->`master`) to respond to the comments in this issue once I check all the items above. Thank you so much, @mfroeling!
