### 2020-07-29 8de0641
Improve Data Panel
* Better directory navigation
* Give users the option to remove a data
