+++ src/Common/tools/FileBrowser/BrowserSet.m
@@ -23,6 +23,7 @@ classdef BrowserSet
     properties(Hidden = true)
         NameText;
         BrowseBtn;
+        ClearBtn;
         InfoBtn;
         FileBox;
         ViewBtn;
@@ -66,12 +67,20 @@ classdef BrowserSet
                 end
 
                 % add Browse button
-                Position = [Location + [0.14, 0], 0.1, 0.1];
+                Position = [Location + [0.14, 0], 0.05, 0.11];
                 obj.BrowseBtn = uicontrol(obj.parent, 'Style', 'pushbutton', 'units', 'normalized', 'fontunits', 'normalized', ...
-                    'String', 'Browse', 'Position', Position, 'FontSize', 0.6,'Interruptible','off');
+                    'String', '', 'Position', Position, 'FontSize', 0.6,'Interruptible','off');
+                im = imread('./src/Common/icons/plus.png');
+                obj.BrowseBtn.CData = im;
+                
+                Position = [Location + [0.19, 0], 0.05, 0.11];
+                obj.ClearBtn = uicontrol(obj.parent, 'Style', 'pushbutton', 'units', 'normalized', 'fontunits', 'normalized', ...
+                    'String', '', 'Position', Position, 'FontSize', 0.6,'Interruptible','off');
+                im = imread('./src/Common/icons/minus.png');
+                obj.ClearBtn.CData = im;
 
                 % add Browse button
-                Position = [Location + [0.27, 0], 0.58, 0.1];
+                Position = [Location + [0.25, 0], 0.58, 0.1];
                 obj.FileBox = uicontrol(obj.parent, 'Style', 'text','units', 'normalized', 'fontunits', 'normalized', 'Position', Position,'FontSize', 0.6,...
                     'BackgroundColor', [1 1 1]);
                 
@@ -85,8 +94,10 @@ classdef BrowserSet
                     'String', 'View', 'Position', Position, 'FontSize', 0.6,'Interruptible','off');
 
                 % Set Callbacks
-                set(obj.FileBox,'Callback', {@(src, event)BrowserSet.BrowseBtn_callback(obj)});
-                set(obj.BrowseBtn,'Callback', {@(src, event)BrowserSet.BrowseBtn_callback(obj)});
+                set(obj.FileBox,'Callback', {@(src, event)BrowserSet.BrowseBtn_callback(obj,info,InputOptional)});
+                set(obj.FileBox,'Callback', {@(src, event)BrowserSet.ClearBtn_callback(obj,info,InputOptional)});
+                set(obj.BrowseBtn,'Callback', {@(src, event)BrowserSet.BrowseBtn_callback(obj,info,InputOptional)});
+                set(obj.ClearBtn,'Callback', {@(src, event)BrowserSet.ClearBtn_callback(obj,info,InputOptional)});
                 set(obj.ViewBtn,'Callback', {@(src, event)BrowserSet.ViewBtn_callback(obj, src, event)});
                 
                 if strcmp(InputName,'Mask')
@@ -104,6 +115,7 @@ classdef BrowserSet
         function Visible(obj, Visibility)
             set(obj.NameText, 'Visible', Visibility);
             set(obj.BrowseBtn, 'Visible', Visibility);
+            set(obj.ClearBtn, 'Visible', Visibility);
             set(obj.FileBox, 'Visible', Visibility);
             if ~strcmp(obj.NameID{1},'Mask')
                 set(obj.ViewBtn, 'Visible', Visibility);
@@ -154,7 +166,9 @@ classdef BrowserSet
                     end
                     tmp = File;
                 else
[…]
