+++ docs/source/gui-example.rst
@@ -46,13 +46,13 @@ Click *Download Data* (blue) button located at the upper right corner of the dat
    :scale: 100 %
 
 You will be prompted for a directory where the example dataset will be saved. After the dataset has been downloaded, qMRLab will 
-automatically set *Path data* to the download directory and load the input files. *Required* input labels are displayed in bold (e.g., `VFAData`), whereas the labels of the *Optional* inputs are in faded color (e.g., `B1map` and `Mask`). You can get more information about the input fields by clicking question mark buttons next to the input labels.
+automatically set *Path data* to the download directory and load the input files.
 
 :: 
 
-  Please note that such auto-loading takes effect only if the name of the images (e.g. VFAData.nii.gz) in the Path data directory are identical with that of the data fields (e.g., VFAData) listed in the data panel.
+  Please note that such auto-loading takes effect only if the name of the images (e.g. VFAData.nii.gz) in the Path data directory are identical to that of the data fields (e.g., VFAData) listed in the data panel. There are no filename assumptions for the user data. Any file name is acceptable as long as the data format and the data dimensions are in agreement with the method's inputs. 
 
-For a more detailed description of what the input data should look like, please refer to :ref:`4.1	Data format`
+For a more detailed description about input data formats, please refer to :ref:`4.1	Data format`
 
 6. View the data
 -------------------------
