+++ docs/source/example_datasets.rst
@@ -15,14 +15,13 @@ Download example datasets via GUI
 After selecting a model from the dropdown menu, you can use *Download Data* (blue) button located at the upper right corner of the data panel to download the respective dataset:
 
 .. figure:: _static/gui_download.png
-   :scale: 100 %
+   :width: 800px
+   :height: 158px
 
 You will be prompted for a directory where the example dataset will be saved. After the dataset has been downloaded, qMRLab will 
 automatically set *Path data* to the download directory and load the input files.
 
-:: 
-
-  Please note that such auto-loading takes effect only if the name of the images (e.g. VFAData.nii.gz) in the Path data directory are identical with that of the data fields (e.g., VFAData) listed in the data panel.
+*Please note that* such auto-loading takes effect only if the name of the images (e.g. VFAData.nii.gz) in the Path data directory are identical with that of the data fields (e.g., VFAData) listed in the data panel.
 
 Step by step instructions to get started with GUI are available at :ref:`Beginner's example with GUI`.
 
