+++ docs/source/gui-example.rst
@@ -43,14 +43,17 @@ For example, in this test dataset, 1 slice at 2 different flip angles were acqui
 Click *Download Data* (blue) button located at the upper right corner of the data panel:
 
 .. figure:: _static/gui_download.png
-   :scale: 100 %
+   :width: 800px
+   :height: 158px
 
 You will be prompted for a directory where the example dataset will be saved. After the dataset has been downloaded, qMRLab will 
 automatically set *Path data* to the download directory and load the input files.
 
-:: 
+.. note::
+   **Please note that** such auto-loading takes effect only if the name of the images (e.g. VFAData.nii.gz) in the Path data directory are identical to that of the data fields (e.g., VFAData) listed in the data panel. There are no filename assumptions for the user data. Any file name is acceptable as long as the data format and the data dimensions are in agreement with the method's inputs. 
 
-  Please note that such auto-loading takes effect only if the name of the images (e.g. VFAData.nii.gz) in the Path data directory are identical to that of the data fields (e.g., VFAData) listed in the data panel. There are no filename assumptions for the user data. Any file name is acceptable as long as the data format and the data dimensions are in agreement with the method's inputs. 
+.. warning::
+   **The default `Protocol` parameters are intended for the example datasets.** Users are required to update these values with regards to their data. 
 
 For a more detailed description about input data formats, please refer to :ref:`4.1	Data format`
 
