### 2020-08-05 #2343 @Kevin-Mattheus-Moerman
@agahkarakuzu can you update on progress? Have the DOI issues been resolved? Have you commented on/resolved all reviewer issues? Feel free to regenerate the pdf (`@whedon generate pdf`) here and also to check for references (`@whedon check references`).
