### 2020-08-05 e5e7333
Various bug fixes
