### 2020-08-05 #2343 @agahkarakuzu
@Kevin-Mattheus-Moerman I resolved all the reviewer issues, just pushed a bug fix, waiting for CI tests, then going to merge `joss_rev2` branch into the `master` today and follow your instructions.
