### 2020-08-06 7def5bc
JOSS Review #2  (#400)
* ADD Missing DOIs

* MATLAB Toolbox Check

- This was available for CLI (startup.m)
- Now also available for GUI

* Improve doc & GUI help

* Supplementary information re example datsets
* Add help button for prots that appears when tabletip is passed 
* Add tabletip for all the diffusion models.

* Make Optional/Required more apparent

* Improve req/opt conditions and input types

* Update B1map definitions 
* Make info appear in the textbox

* R1map was made required 42acaa3

* Patch amico for retrocompatibility

* Improve Data Panel

* Better directory navigation
* G
[…]
