+++ paper.bib
@@ -110,7 +110,8 @@
   number={4},
   pages={316--319},
   year={2017},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nbt.3820}
 }
 
 @article{theaud:2020,
@@ -119,7 +120,8 @@
   journal={NeuroImage},
   pages={116889},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={oi.org/10.1101/631952}
 }
 
 @inproceedings{cook:2006,
@@ -139,7 +141,8 @@
   volume={8},
   pages={8},
   year={2014},
-  publisher={Frontiers}
+  publisher={Frontiers},
+  doi={doi.org/10.3389/fninf.2014.00008}
 }
 
 @article{fick:2019,
@@ -149,5 +152,6 @@
   volume={13},
   pages={64},
   year={2019},
-  publisher={Frontiers}
+  publisher={Frontiers},
+  doi={doi.org/10.3389/fninf.2019.00064}
 }
