### 2020-08-20 #2343 @62442katieb
> @62442katieb thanks again for your help here. You have some boxes that remain unticked. Can you summarize what the authors can do to resolve any remaining issues you have? Thanks!

Hi @Kevin-Mattheus-Moerman ! I've had a couple issues with installation on our cluster. Those issues have been resolved, but now it appears the download links in the [installation instructions](https://qmrlab.readthedocs.io/en/master/install.html) are broken (https://github.com/qMRLab/qMRLab/releases/latest/download/qMRLab_latest.zip, https://github.com/qMRLab/qMRLab/releases/download/v2.4.0/qMRLab_latest.zip).

Please let me know if there is another location from which I should be downloading. Thanks!
