### 2020-08-20 #2343 @agahkarakuzu
@62442katieb can you download the `v2.4.0` from the assets section of [this link](https://github.com/qMRLab/qMRLab/releases/tag/v2.4.0)?

You can also check out the [master](https://github.com/qMRLab/qMRLab/archive/master.zip) version, where previous comments were addressed.

I'll fix the broken hyperlinks.
