### 2020-09-02 #2343 @Kevin-Mattheus-Moerman
@agahkarakuzu I've reviewed your paper and it seems fine. Below are the last steps to complete:

- [ ] Can you archive a copy of the software on ZENODO and report back here with the DOI of the archived version (do not alter or set the DOI yourself, please accept the ZENODO generated DOI)? Please ensure that the title and author list is the same as the JOSS paper for this ZENODO archived version (you may need to edit this meta-data manually).

- [ ] Can you inform me about the version tag for this work. Is it still at `v2.4.0` or has it moved on?

- [ ] Your paper is about to be accepted in JOSS. I recommend that you review the paper yourself once more, and that you also check the author list, affiliations, and the acknowledgements section for accuracy. (call `@whedon generate pdf` here to update the paper).
