### 2020-09-02 #2343 @agahkarakuzu
@Kevin-Mattheus-Moerman thanks! Now that the review has been completed, I would like to release `v.2.4.1` that includes changes suggested by the reviewers and update the citation information to the JOSS reference. Is there a way to know how the citation is going to look like at this stage?
