### 2020-09-02 #2343 @Kevin-Mattheus-Moerman
@agahkarakuzu That is fine, I can set `v.2.4.1` as the version once you confirm that. That way it will be recorded with the paper publication meta data.

@openjournals/dev @arfon I take it that the DOI for this submission will become `https://doi.org/10.21105/joss.02343` (as currently shown on the proof)? Perhaps @agahkarakuzu can use that with the citation information for this release.
