### 2020-09-02 #2343 @arfon
Yes, `https://doi.org/10.21105/joss.02343` will be the DOI.
