### 2020-09-02 71fafd6
v2.4.1 minor changes
- Update CHANGELOG.md 
- Change citation reference
