+++ src/Models/Processing/filter_map.m
@@ -25,11 +25,9 @@ classdef filter_map < AbstractModel & FilterClass
     %
     % References:
     %   Please cite the following if you use this module:
-    %     Cabana J-F, Gu Y, Boudreau M, Levesque IR, Atchia Y, Sled JG,
-    %     Narayanan S, Arnold DL, Pike GB, Cohen-Adad J, Duval T, Vuong M-T and
-    %     Stikov N. (2016), Quantitative magnetization transfer imaging made
-    %     easy with qMTLab: Software for data simulation, analysis, and
-    %     visualization. Concepts Magn. Reson.. doi: 10.1002/cmr.a.21357
+    %     Karakuzu A., Boudreau M., Duval T.,Boshkovski T., Leppert I.R., Cabana J.F., 
+    %     Gagnon I., Beliveau P., Pike G.B., Cohen-Adad J., Stikov N. (2020), qMRLab: 
+    %     Quantitative MRI analysis, under one umbrella doi: 10.21105/joss.02343
     properties (Hidden=true)
         onlineData_url = 'https://osf.io/d8p4h/download?version=1';
     end
