+++ src/Models/T1_relaxometry/vfa_t1.m
@@ -42,11 +42,9 @@ classdef vfa_t1 < AbstractModel
 %     T1 using variable flip angle gradient refocused imaging. Magn. Reson.
 %     Imaging 5, 201?208
 %   In addition to citing the package:
-%     Cabana J-F, Gu Y, Boudreau M, Levesque IR, Atchia Y, Sled JG,
-%     Narayanan S, Arnold DL, Pike GB, Cohen-Adad J, Duval T, Vuong M-T and
-%     Stikov N. (2016), Quantitative magnetization transfer imaging made
-%     easy with qMTLab: Software for data simulation, analysis, and
-%     visualization. Concepts Magn. Reson.. doi: 10.1002/cmr.a.21357
+%     Karakuzu A., Boudreau M., Duval T.,Boshkovski T., Leppert I.R., Cabana J.F., 
+%     Gagnon I., Beliveau P., Pike G.B., Cohen-Adad J., Stikov N. (2020), qMRLab: 
+%     Quantitative MRI analysis, under one umbrella doi: 10.21105/joss.02343
 
 properties (Hidden=true)
  onlineData_url = 'https://osf.io/7wcvh/download?version=3';  
