### 2020-09-02 2f72f78
V2.4.1 (#405)
* v2.4.1 minor changes 

- Update CHANGELOG.md 
- Change citation reference

* Update DOI

* Fix badge
