### 2020-09-02 #2343 @agahkarakuzu
@Kevin-Mattheus-Moerman

> Can you archive a copy of the software on ZENODO and report back here with the DOI of the archived version (do not alter or set the DOI yourself, please accept the ZENODO generated DOI)? Please ensure that the title and author list is the same as the JOSS paper for this ZENODO archived version (you may need to edit this meta-data manually).

**Zenodo DOI:** 10.5281/zenodo.4012104

[![DOI](https://zenodo.org/badge/41506224.svg)](https://zenodo.org/badge/latestdoi/41506224)

✅ Author list on Zenodo matches that on the PDF.

>  Can you inform me about the version tag for this work. Is it still at v2.4.0 or has it moved on?

✅ Tag is now at [`v2.4.1`](https://github.com/qMRLab/qMRLab/releases/tag/v2.4.1).

The [`v2.4.1.` release pipeline](https://dev.azure.com/neuropoly/qMRLab/_releaseProgress?_a=release-pipeline-progress&releaseId=56) has been succeeded. Compiled (MCR) version of the `v2.4.1` is pushed to [DockerHub](https://hub.docker.com/u/qmrlab) along with other virtual environments.

> Your paper is about to be accepted in JOSS. I recommend that you review the paper yourself once more, and that you also check the author list, affiliations, and the acknowledgements section for accuracy.

✅ Author list, affiliations and the acknowledgements are OK.
