### 2020-09-02 #2343 @Kevin-Mattheus-Moerman
@agahkarakuzu can you amend the title of the ZENODO archived version to match that of the paper? It is currently `qMRLab/qMRLab: Release v2.4.1 Zenodo Archive` but should be `qMRLab: Quantitative MRI analysis, under one umbrella`
