+++ paper.bib
@@ -21,7 +21,7 @@
 }
 
 @article{tabelow:2019,
-  title={hMRI--A toolbox for quantitative MRI in neuroscience and clinical research},
+  title={{hMRI}--A toolbox for quantitative {MRI} in neuroscience and clinical research},
   author={Tabelow, Karsten and Balteau, Evelyne and Ashburner, John and Callaghan, Martina F and Draganski, Bogdan and Helms, Gunther and Kherif, Ferath and Leutritz, Tobias and Lutti, Antoine and Phillips, Christophe and others},
   journal={Neuroimage},
   volume={194},
@@ -52,7 +52,7 @@
 }
 
 @article{hafyane:2018,
-  title={Let's talk about cardiac T1 mapping},
+  title={Let's talk about cardiac {T1} mapping},
   author={{Hafyane}, T. and {Karakuzu}, A. and {Duquette}, C. and {Mongeon}, F.~P. and {Cohen-Adad}, J. and {Herold}, M.~J. and {Friedrich}, M.~G. and {Stikov}, N.},
   journal={BioRxiv},
   pages={343079},
@@ -62,7 +62,7 @@
 }
 
 @article{werys:2020,
-  title={Total Mapping Toolbox (TOMATO): An open source library for cardiac magnetic resonance parametric mapping},
+  title={Total Mapping Toolbox {(TOMATO)}: An open source library for cardiac magnetic resonance parametric mapping},
   author={{Werys}, K. and {Dragonu}, I. and {Zhang}, Q. and {Popescu}, I. and {Hann}, E. and {Puchta}, H. and {Kubik}, A. and {Polat}, D. and {Wu}, C. and {Moon}, N.~O.},
   journal={SoftwareX},
   volume={11},
@@ -83,7 +83,7 @@
 
 @misc{samsonov:2011,
   author = {{Hurley}, S.~A. and {Mossahebi}, P. and {Samsonov}, A.~A.},
-  title = {Quantitative MRI Analysis Package},
+  title = {Quantitative {MRI} Analysis Package},
   year = {2011},
   publisher = {University of Wisconsin - Madison},
   journal = {Code repository},
@@ -115,7 +115,7 @@
 }
 
 @article{theaud:2020,
-  title={TractoFlow: A robust, efficient and reproducible diffusion MRI pipeline leveraging Nextflow \& Singularity},
+  title={TractoFlow: A robust, efficient and reproducible diffusion {MRI} pipeline leveraging {Nextflow} \& {Singularity}},
   author={Theaud, Guillaume and Houde, Jean-Christophe and Bor{\'e}, Arnaud and Rheault, Fran{\c{c}}ois and Morency, Felix and Descoteaux, Maxime},
   journal={NeuroImage},
   pages={116889},
@@ -125,7 +125,7 @@
 }
 
 @inproceedings{cook:2006,
-  title={Camino: open-source diffusion-MRI reconstruction and processing},
+  title={Camino: {open-source diffusion-MRI} reconstruction and processing},
   author={Cook, PA and Bai, Yu and Nedjati-Gilani, SKKS and Seunarine, KK and Hall, MG and Parker, GJ and Alexander, D Cq},
   booktitle={14th scientific meeting of the international society for magnetic resonance in medicine},
   volume={2759},
@@ -135,7 +135,7 @@
 }
 
 @article{garyfallidis:2014,
-  title={Dipy, a library for the analysis of diffusion MRI data},
+  title={Dipy, a library for the analysis of diffusion {MRI} data},
   author={Garyfallidis, Eleftherios and Brett, Matthew and Amirbekian, Bagrat and Rokem, Ariel and Van Der Walt, Stefan and Descoteaux, Maxime and Nimmo-Smith, Ian},
   journal={Frontiers in neuroinformatics},
   volume={8},
