### 2020-09-03 #2343 @kthyng
Hi @agahkarakuzu, I am the rotating EIC and I just read through your paper and I see that the capitalization in your references is incorrect. I made changes to your .bib file in [this PR](https://github.com/qMRLab/qMRLab/pull/407).
